refactor some tests - add some robustness unittests
