--- android/sdk/src/test/java/io/o2mc/sdk/ExampleUnitTest.java
@@ -1,18 +1,0 @@
-package io.o2mc.sdk;
-
-import org.junit.Test;
-
-import static org.junit.Assert.assertEquals;
-
-/**
- * Example local unit test, which will execute on the development machine (host).
- *
- * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
- */
-public class ExampleUnitTest {
-    @SuppressWarnings("RedundantThrows")
-    @Test
-    public void addition_isCorrect() throws Exception {
-        assertEquals(4, 2 + 2);
-    }
-}
